refactor(forge): extract local smoke seed helpers for maintainability
Move seed logic into lib/forge-local-smoke/seed-task.js so functions stay
under the 50-line maintainability cap. Map the new module in change
ownership and point unit tests at the lib entrypoint.
