+++ config/change-ownership-map.json
@@ -490,7 +490,8 @@
         "^db/migrations/010_merge_readiness_reviews\\.sql$",
         "^db/migrations/012_projects(\\.down)?\\.sql$",
         "^db/migrations/014_ai_agents_management\\.sql$",
-        "^scripts/(backfill-task-platform|rollout-task-platform|verify-task-platform-rollout|verify-merge-readiness-branch-protection|emit-merge-readiness-check|verify-projects-production-smoke)\\.js$"
+        "^lib/forge-local-smoke/seed-task\\.js$",
+        "^scripts/(backfill-task-platform|rollout-task-platform|verify-task-platform-rollout|verify-merge-readiness-branch-protection|emit-merge-readiness-check|verify-projects-production-smoke|seed-forge-local-smoke-task)\\.js$"
       ],
       "test_requirements": [
         {
@@ -505,7 +506,8 @@
         "^tests/unit/task-platform-pr-summary\\.test\\.js$",
         "^tests/unit/task-platform-merge-readiness-gate\\.test\\.js$",
         "^tests/unit/projects-production-smoke\\.test\\.js$",
-        "^tests/unit/forge-canonical-task\\.test\\.js$"
+        "^tests/unit/forge-canonical-task\\.test\\.js$",
+        "^tests/unit/seed-forge-local-smoke-task\\.test\\.js$"
       ]
     },
         {
@@ -532,6 +534,7 @@
           "name": "runbook-or-adr",
           "patterns": [
             "^docs/runbooks/task-platform-rollout\\.md$",
+            "^docs/runbooks/forge-local-smoke\\.md$",
             "^docs/adr/ADD-2026-04-12-task-persistence-and-agent-ownership\\.md$"
           ]
         }
